fix shortcut session feedback
