+++ data/create-shortcuts-core.ts
@@ -1,6 +1,6 @@
 import type { Shortcut, ShortcutDescription } from '@/types/interfaces'
 
-import { createShortcut } from './create-shortcuts'
+import { createShortcut } from './create-shortcut'
 
 export const createShortcutsFromRecords = (
   records: ShortcutDescription[],
